Move code under prepare
